requirements.txt
streamlit
google-generativeai
langchain
langchain-google-genai
langchain-community
faiss-cpu
pypdf
python-dotenv
PyPDF2
sentence-transformers
pymupdf